trivial: rename chromeos to chromiumos
